advanced_rag/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.25.0
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
numpy==1.26.4
httpx==0.27.0
websockets==12.0
python-multipart==0.0.9
prometheus-fastapi-instrumentator==6.1.0

# Ingestion/Parsing
pypdf==4.2.0
python-docx==1.1.2
pandas==2.2.2
openpyxl==3.1.5
pytesseract==0.3.10
Pillow==10.4.0
pdf2image==1.17.0
camelot-py[cv]==0.11.0
tabula-py==2.9.0
detectron2 @ git+https://github.com/facebookresearch/detectron2.git
langdetect==1.0.9
spacy==3.7.2
transformers==4.36.0
torch==2.2.0
# Retrieval and ML
faiss-cpu==1.8.0.post1
scikit-learn==1.5.1
sentence-transformers==3.0.1
rank-bm25==0.2.2

# Providers
openai==1.40.1
cohere==5.5.8
anthropic==0.34.2

# Testing and templates
pytest==8.3.2
Jinja2==3.1.4
chainlit==1.2.0
aiohttp==3.9.5

# Security and authentication
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
PyJWT==2.8.0

# Additional dependencies for new features
psutil==5.9.5

# Ollama and local model support
ollama==0.3.1

# Enhanced image processing
opencv-python==4.8.1.78
Pillow==10.4.0

# Math and code processing
sympy==1.12
matplotlib==3.8.2
pygments==2.17.2

# Translation services
googletrans==3.1.0a0
python-dateutil==2.8.2


